Re-organized folder structures. Created src file.
