+++ src/containers/ForeCast/ForeCast.module.css
@@ -1,5 +1,5 @@
 .forecastContainer {
-    background-image: url('../../public/images/daytime-background.jpg');
+    background-image: url('../../../public/images/daytime-background.jpg');
     /* background-image: linear-gradient( to bottom, rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.75)); */
     background-repeat: no-repeat;
     background-size: cover;
